model_training/neuralop_requirements.txt
wandb
ruamel.yaml
configmypy
zencfg
tensorly
tensorly-torch
torch-harmonics
matplotlib
opt-einsum
h5py
zarr